Move tpu client and varieties to new crate
