+++ client/Cargo.toml
@@ -48,6 +48,7 @@ solana-net-utils = { path = "../net-utils", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
 solana-streamer = { path = "../streamer", version = "=1.12.0" }
 solana-thin-client = { path = "../thin-client", version = "=1.12.0" }
+solana-tpu-client = { path = "../tpu-client", version = "=1.12.0" }
 solana-transaction-status = { path = "../transaction-status", version = "=1.12.0" }
 solana-version = { path = "../version", version = "=1.12.0" }
 solana-vote-program = { path = "../programs/vote", version = "=1.12.0" }
