+++ thin-client/Cargo.toml
@@ -14,6 +14,7 @@ bincode = "1.3.3"
 log = "0.4.17"
 solana-client = { path = "../client", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
+solana-tpu-client = { path = "../tpu-client", version = "=1.12.0" }
 
 [dev-dependencies]
 rayon = "1.5.3"
