+++ tpu-client/Cargo.toml
@@ -1,0 +1,45 @@
+[package]
+name = "solana-tpu-client"
+version = "1.12.0"
+description = "Solana TPU Client"
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
+repository = "https://github.com/solana-labs/solana"
+license = "Apache-2.0"
+homepage = "https://solana.com/"
+documentation = "https://docs.rs/solana-tpu-client"
+edition = "2021"
+
+[dependencies]
+async-mutex = "1.4.0"
+async-trait = "0.1.57"
+bincode = "1.3.3"
+enum_dispatch = "0.3.8"
+futures = "0.3"
+futures-util = "0.3.21"
+indexmap = "1.9.1"
+itertools = "0.10.2"
+lazy_static = "1.4.0"
+log = "0.4.17"
+quinn = "0.8.4"
+quinn-proto = "0.8.4"
+quinn-udp = "0.1.3"
+rand = "0.7.0"
+rayon = "1.5.3"
+rustls = { version = "0.20.6", features = ["dangerous_configuration"] }
+solana-client = { path = "../client", version = "=1.12.0" }
+solana-measure = { path = "../measure", version = "=1.12.0" }
+solana-metrics = { path = "../metrics", version = "=1.12.0" }
+solana-net-utils = { path = "../net-utils", version = "=1.12.0" }
+solana-sdk = { path = "../sdk", version = "=1.12.0" }
+solana-streamer = { path = "../streamer", version = "=1.12.0" }
+thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
+
+[dev-dependencies]
+crossbeam-channel = "0.5"
+rand_chacha = "0.2.2"
+solana-logger = { path = "../logger", version = "=1.12.0" }
+solana-perf = { path = "../perf", version = "=1.12.0" }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
