+++ tpu-client/src/lib.rs
@@ -1,0 +1,11 @@
+#![allow(clippy::integer_arithmetic)]
+
+pub mod connection_cache;
+pub mod nonblocking;
+pub mod quic_client;
+pub mod tpu_client;
+pub mod tpu_connection;
+pub mod udp_client;
+
+#[macro_use]
+extern crate solana_metrics;
